Conflicts:
	ide
	libfoundation/include/foundation.h
	thirdparty
	version

+++ engine/src/dispatch.h
@@ -207,7 +207,7 @@ public:
 	MCStack *findstackname(MCNameRef);
 	MCStack *findstackid(uint4 fid);
 	// IM-2014-07-09: [[ Bug 12225 ]] Find the stack by window ID
-	MCStack *findstackwindowid(uint32_t p_win_id);
+	MCStack *findstackwindowid(uintptr_t p_win_id);
 	MCStack *findstackd(Window w);
 	
 	// IM-2014-07-23: [[ Bug 12930 ]] Replace findchildstack method with iterating method
